fixed GetTime

+++ Tabby/src/Tabby/Core/Timestep.h
@@ -4,23 +4,23 @@ namespace Tabby {
 
 class Timestep {
 public:
-    Timestep(float time = 0.0f)
+    Timestep(double time = 0.0f)
         : m_Time(time)
     {
     }
 
-    operator float() const
+    operator double() const
     {
 
         return m_Time;
     }
 
-    float GetSeconds() const { return m_Time; }
-    float GetMilliseconds() const { return m_Time * 1000.0f; }
+    double GetSeconds() const { return m_Time; }
+    double GetMilliseconds() const { return m_Time * 1000.0f; }
 
 private:
     // static Timestep* s_Instance;
-    float m_Time;
+    double m_Time;
 };
 
 }
